refactor : 패키지 이름 scrap 에서 storage 로 수정
